refactor(proxy): share deferred cancellation helper
